requirements.txt
Flask==2.1.1
Flask-Cors==3.0.10
gunicorn==20.1.0
Pillow==9.1.1
numpy==1.22.3
pandas==1.4.2
scikit-learn==1.0.2
Werkzeug==2.3.8

# Optimized for Render (Linux) deployment
tensorflow-cpu==2.8.0
tflite-runtime==2.8.0